Add test to set the GraceTime of a contianer dynamically
[#101620462]

+++ garden_integration_tests_suite_test.go
@@ -50,7 +50,9 @@ func TestGardenIntegrationTests(t *testing.T) {
 	})
 
 	AfterEach(func() {
-		Expect(gardenClient.Destroy(container.Handle())).To(Succeed())
+		if container != nil {
+			Expect(gardenClient.Destroy(container.Handle())).To(Succeed())
+		}
 	})
 
 	RunSpecs(t, "GardenIntegrationTests Suite")
